+++ transmonee_dashboard/src/transmonee_dashboard/pages/base_page.py
@@ -17,16 +17,16 @@ from ..components import fa
 from . import (
     countries,
     countries_iso3_dict,
-    df_sources,
+    # df_sources,
     dimension_names,
     geo_json_countries,
     get_filtered_dataset,
     indicator_names,
     indicators_config,
-    programme_country_indexes,
+    # programme_country_indexes,
     selection_index,
     selection_tree,
-    unicef_country_prog,
+    # unicef_country_prog,
     years,
     get_search_countries,
 )
@@ -51,6 +51,7 @@ DEFAULT_LABELS = {
     "Country_name": "Country",
     "TIME_PERIOD": "Year",
     "Sex_name": "Sex",
+    "Educationlevel_name": "Education level",
     "Residence_name": "Residence",
     "Age_name": "Age",
     "Wealth_name": "Wealth Quintile",
@@ -304,24 +305,24 @@ def get_base_layout(**kwargs):
                                             ),
                                         ],
                                     ),
-                                    dbc.FormGroup(
-                                        [
-                                            dbc.Checkbox(
-                                                id="programme-toggle",
-                                                className="custom-control-input",
-                                            ),
-                                            dbc.Label(
-                                                "UNICEF Country Programmes",
-                                                html_for="programme-toggle",
-                                                className="custom-control-label",
-                                                color="primary",
-                                            ),
-                                        ],
-                                        className="custom-control custom-switch m-2",
-                                        check=True,
-                                        inline=True,
-                                        style=programme_toggle_style,
-                                    ),
+                                    # dbc.FormGroup(
+                                    #     [
+                                    #         dbc.Checkbox(
+                                    #             id="programme-toggle",
+                                    #             className="custom-control-input",
+                                    #         ),
+                                    #         dbc.Label(
+                                    #             "UNICEF Country Programmes",
+                                    #             html_for="programme-toggle",
+                                    #             className="custom-control-label",
+                                    #             color="primary",
+                                    #         ),
+                                    #     ],
+                                    #     className="custom-control custom-switch m-2",
+                                    #     check=True,
+                                    #     inline=True,
+                                    #     style=programme_toggle_style,
+                                    # ),
                                 ],
                                 id="filter-row",
                                 no_gutters=True,
@@ -527,29 +528,37 @@ def apply_filters(
     # check if it is the country profile page
     is_country_profile = current_theme == "COUNTRYPROFILE"
     # check if the user clicked on the generate button in the country profile page
-    if is_country_profile:
-        key_list = list(countries_iso3_dict.keys())
-        val_list = list(countries_iso3_dict.values())
-        # get the name of the selected country in the dropdown to filter the data accordingly
-        countries_selected = (
-            [key_list[val_list.index(selected_country)]] if selected_country else []
-        )
-    elif programme_toggle and selected == "programme-toggle":
-        countries_selected = unicef_country_prog
-        country_selector = programme_country_indexes
-    # Add the condition to know when the user unchecks the UNICEF country programs!
-    elif not country_selector or (
-        not programme_toggle and selected == "programme-toggle"
-    ):
-        countries_selected = countries
-        # Add this to check all the items in the selection tree
-        country_selector = ["0"]
-    else:
-        for index in country_selector:
-            countries_selected.update(selection_index[index])
-            if countries_selected == countries:
-                # if all countries are all selected then stop
-                break
+
+
+    # if is_country_profile:
+    #     key_list = list(countries_iso3_dict.keys())
+    #     val_list = list(countries_iso3_dict.values())
+    #     # get the name of the selected country in the dropdown to filter the data accordingly
+    #     countries_selected = (
+    #         [key_list[val_list.index(selected_country)]] if selected_country else []
+    #     )
+    # elif programme_toggle and selected == "programme-toggle":
+    #     countries_selected = unicef_country_prog
+    #     country_selector = programme_country_indexes
+    # # Add the condition to know when the user unchecks the UNICEF country programs!
+    # elif not country_selector or (
+    #     not programme_toggle and selected == "programme-toggle"
+    # ):
+    #     countries_selected = countries
+    #     # Add this to check all the items in the selection tree
+    #     country_selector = ["0"]
+    # else:
+    #     for index in country_selector:
+    #         countries_selected.update(selection_index[index])
+    #         if countries_selected == countries:
+    #             # if all countries are all selected then stop
+    #             break
+
+    for index in country_selector:
+        countries_selected.update(selection_index[index])
+        if countries_selected == countries:
+            # if all countries are all selected then stop
+            break
 
     countries_selected = list(countries_selected)
     country_text = f"{len(countries_selected)} Selected"
@@ -612,18 +621,20 @@ def indicator_card(
         latest_data=True,
     )
 
-    df_indicator_sources = df_sources[df_sources["Code"].isin(indicators)]
-    unique_indicator_sources = df_indicator_sources["Source_Full"].unique()
+    # df_indicator_sources = df_sources[df_sources["Code"].isin(indicators)]
+    # unique_indicator_sources = df_indicator_sources["Source_Full"].unique()
+    unique_indicator_sources = []
     indicator_sources = (
         "; ".join(list(unique_indicator_sources))
         if len(unique_indicator_sources) > 0
         else ""
     )
-    source_link = (
-        df_indicator_sources["Source_Link"].unique()[0]
-        if len(unique_indicator_sources) > 0
-        else ""
-    )
+    # source_link = (
+    #     df_indicator_sources["Source_Link"].unique()[0]
+    #     if len(unique_indicator_sources) > 0
+    #     else ""
+    # )
+    source_link = ""
     # lbassil: add this check because we are getting an exception where there is no data; i.e. no totals for all dimensions mostly age for the selected indicator
     if filtered_data.empty:
         indicator_header = "No data"
@@ -931,18 +942,20 @@ def main_figure(indicator, show_historical_data, selections, indicators_dict):
         if len(data[data["CODE"] == indicator]["Unit_name"].astype(str).unique()) > 0
         else ""
     )
-    df_indicator_sources = df_sources[df_sources["Code"] == indicator]
-    unique_indicator_sources = df_indicator_sources["Source_Full"].unique()
+    # df_indicator_sources = df_sources[df_sources["Code"] == indicator]
+    # unique_indicator_sources = df_indicator_sources["Source_Full"].unique()
     source = (
         "; ".join(list(unique_indicator_sources))
         if len(unique_indicator_sources) > 0
         else ""
     )
-    source_link = (
-        df_indicator_sources["Source_Link"].unique()[0]
-        if len(unique_indicator_sources) > 0
-        else ""
-    )
+    source = ""
+    # source_link = (
+    #     df_indicator_sources["Source_Link"].unique()[0]
+    #     if len(unique_indicator_sources) > 0
+    #     else ""
+    # )
+    source_link = ""
 
     options["labels"] = DEFAULT_LABELS.copy()
     options["labels"]["OBS_VALUE"] = name
