--- transmonee_dashboard/src/transmonee_dashboard/pages/country_profiles.py
@@ -1,601 +1,0 @@
-import dash
-from dash_bootstrap_components._components.Row import Row
-import dash_core_components as dcc
-import dash_html_components as html
-import dash_bootstrap_components as dbc
-from dash_html_components.Br import Br
-import dash_table
-from dash.dependencies import Input, State, Output
-import pandas as pd
-import plotly.express as px
-
-from ..app import app
-from . import (
-    get_search_countries,
-    countries_iso3_dict,
-    data,
-    df_topics_subtopics,
-    topics_subtopics,
-)
-
-from .base_page import get_base_layout, geo_json_countries
-
-indicators_dict = {
-    "COUNTRYPROFILE": {
-        "NAME": "Country Name",
-        "CARDS": [
-            {
-                "name": "Age 0-4 years",
-                "indicator": "DM_CHLD_POP",
-                "suffix": "Total child population (expressed in thousands)",
-                "absolute": True,
-                "age": "Y0T4",  # Under 5 years old
-            },
-            {
-                "name": "Age 10-19 years",
-                "indicator": "DM_ADOL_POP",
-                "suffix": "Total adolescent population (expressed in thousands)",
-                "absolute": True,
-                "age": "Y10T19",
-            },
-            {
-                "name": "age 0-17 years",
-                "indicator": "DM_CHLD_POP",
-                "suffix": "Total child population (expressed in thousands)",
-                "absolute": True,
-                "age": "Y0T17",
-            },
-            {
-                "name": "Children as a share of the total population",
-                "indicator": "DM_CHLD_POP_PT",
-                "suffix": "Percentage of",
-                "min_max": True,
-                "age": "Y0T17",
-            },
-            {
-                "name": "Age 15-24 years",
-                "indicator": "DM_ADOL_YOUTH_POP",
-                "suffix": "Total young population (expressed in thousands)",
-                "absolute": True,
-                "age": "Y15T24",  # 15 to 24 years old
-            },
-        ],
-        "AREA_1": {
-            "name": "Education, Leisure, and Culture",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "default_graph": "line",
-            "indicators": [
-                "EDUNF_ROFST_L1",
-                "EDUNF_ROFST_L2",
-                "EDUNF_ROFST_L3",
-                "EDUNF_OFST_L1",
-                "EDUNF_OFST_L2",
-                "EDUNF_OFST_L3",
-                "EDUNF_ROFST_L1_UNDER1",
-                "EDUNF_ROFST_L1T3",
-                "EDUNF_STU_L1_TOT",
-                "EDUNF_STU_L2_TOT",
-                "EDUNF_STU_L3_TOT",
-                "EDUNF_GER_L2_GEN",
-                "EDUNF_GER_L2_VOC",
-                "EDUNF_GER_L3_GEN",
-                "EDUNF_GER_L3_VOC",
-                "EDUNF_NER_L02",
-                "EDUNF_NERA_L1_UNDER1",
-                "EDUNF_NERA_L1",
-                "EDUNF_NERA_L2",
-                "EDUNF_GER_L1",
-                "EDUNF_GER_L2",
-                "EDUNF_GER_L3",
-                "EDUNF_NIR_L1_ENTRYAGE",
-                "EDUNF_TRANRA_L2",
-                "EDU_SDG_SCH_L1",
-                "EDU_SDG_SCH_L2",
-                "EDU_SDG_SCH_L3",
-                "WS_SCH_H-B",
-                "WS_SCH_S-B",
-                "WS_SCH_W-B",
-                "EDU_CHLD_DISAB",
-                "EDU_CHLD_DISAB_GENERAL",
-                "EDU_CHLD_DISAB_SPECIAL",
-                "EDU_SDG_PRYA",
-                "EDUNF_CR_L1",
-                "EDUNF_CR_L2",
-                "EDUNF_CR_L3",
-                "EDUNF_DR_L1",
-                "EDUNF_DR_L2",
-                "ECD_CHLD_36-59M_LMPSL",
-                "EDU_SDG_STU_L2_GLAST_MAT",
-                "EDU_SDG_STU_L2_GLAST_REA",
-                "EDU_SDG_STU_L1_GLAST_MAT",
-                "EDU_SDG_STU_L1_G2OR3_MAT",
-                "EDU_SDG_STU_L1_GLAST_REA",
-                "EDU_SDG_STU_L1_G2OR3_REA",
-                "EDUNF_LR_YOUTH",
-                "EDUNF_LR_ADULT",
-                "EDU_SDG_TRTP_L02",
-                "EDU_SDG_TRTP_L1",
-                "EDU_SDG_TRTP_L2",
-                "EDU_SDG_TRTP_L3",
-                "EDU_SDG_QUTP_L02",
-                "EDU_SDG_QUTP_L1",
-                "EDU_SDG_QUTP_L2",
-                "EDU_SDG_QUTP_L3",
-                "EDU_FIN_EXP_PT_GDP",
-                "EDU_FIN_EXP_PT_TOT",
-                "EDU_SDG_FREE_EDU_L02",
-                "EDU_SDG_COMP_EDU_L02",
-                "EDU_FIN_EXP_CONST_PPP",
-                "EDUNF_PRP_L02",
-                "EDUNF_PRP_L1",
-                "EDUNF_PRP_L2",
-                "EDUNF_PRP_L3",
-                "EDUNF_STU_L01_PUB",
-                "EDUNF_STU_L02_PUB",
-                "EDUNF_STU_L1_PUB",
-                "EDUNF_STU_L2_PUB",
-                "EDUNF_STU_L3_PUB",
-                "EDUNF_STU_L01_PRV",
-                "EDUNF_STU_L02_PRV",
-                "EDUNF_STU_L1_PRV",
-                "EDUNF_STU_L2_PRV",
-                "EDUNF_STU_L3_PRV",
-                "EDU_FIN_EXP_L02",
-                "EDU_FIN_EXP_L1",
-                "EDU_FIN_EXP_L2",
-                "EDU_FIN_EXP_L3",
-                "EDU_FIN_EXP_L4",
-                "EDU_FIN_EXP_L5T8",
-                "EDU_FIN_EXP_CONST_PPP",
-                "EDUNF_ADMIN_L1_GLAST_REA",
-                "EDUNF_ADMIN_L1_GLAST_MAT",
-                "EDUNF_ADMIN_L2_REA",
-                "EDUNF_ADMIN_L2_MAT",
-                "EDUNF_ADMIN_L1_G2OR3_REA",
-                "EDUNF_ADMIN_L1_G2OR3_MAT",
-            ],
-            "default": "EDU_FIN_EXP_PT_GDP",
-        },
-        "AREA_2": {
-            "name": "Family environment and protection from violence and harmful practices",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "indicators": [
-                "PT_F_GE15_PS-SX-EM_V_PTNR_12MNTH",
-                "PT_F_GE15_SX_V_PTNR_12MNTH",
-                "PT_VC_VOV_PHYL",
-                "PT_VC_VOV_SEXL",
-                "PT_VC_VOV_ROBB",
-                "PT_VC_SNS_WALN",
-                "PT_CHLD_1-14_PS-PSY-V_CGVR",
-                "PT_F_18-29_SX-V_AGE-18",
-                "PT_M_18-29_SX-V_AGE-18",
-                "PT_VC_PRR_PHYV",
-                "PT_VC_PRR_SEXV",
-                "PT_VC_PRR_ROBB",
-                "PT_ADLT_PS_NEC",
-                "PT_F_15-49_W-BTNG",
-                "PT_M_15-49_W-BTNG",
-                "PT_ST_13-15_BUL_30-DYS",
-                "PT_CHLD_INRESIDENTIAL",
-                "PT_CHLD_INRESIDENTIAL_RATE_B",
-                "PT_CHLD_DISAB_PUBLIC",
-                "PT_CHLD_LEFTRESCARE",
-                "PT_CHLD_LEFTRESCARE_RETURNED",
-                "PT_CHLD_LEFTRESCARE_INFAMILY",
-                "PT_CHLD_LEFTRESCARE_ADOPTED",
-                "PT_CHLD_LEFTRESCARE_INDEPENDENT",
-                "PT_CHLD_LEFTRESCARE_TRANSFERED",
-                "PT_CHLD_LEFTRESCARE_DIED",
-                "PT_CHLD_LEFTRESCARE_OTHER",
-                "PT_CHLD_CARED_BY_FOSTER",
-                "PT_CHLD_CARED_BY_FOSTER_RATE",
-                "PT_CHLD_DISAB_FOSTER",
-                "PT_CHLD_CARED_GUARDIAN",
-                "PT_CHLD_CARED_GUARDIAN_RATE",
-                "PT_CHLD_DISAB_CARED_GUARDIAN",
-                "PT_CHLD_GUARDIAN",
-                "PT_CHLD_ENTEREDFOSTER",
-                "PT_CHLD_ADOPTION",
-                "PT_CHLD_ADOPTION_RATE",
-                "PT_CHLD_ADOPTION_DISAB",
-                "PT_CHLD_ADOPTION_AVAILABLE",
-                "PT_CHLD_ADOPTION_AVAILABLE_DISAB",
-                "JJ_CHLD_DETENTION",
-                "JJ_CHLD_CONVICTED",
-                "JJ_CHLD_SENTENCERT",
-                "JJ_VC_PRS_UNSNT",
-                "JJ_PRISIONERS",
-                "JJ_PRISIONERS_RT",
-                "PT_F_15-19_MRD",
-                "PT_M_15-19_MRD",
-                "PT_F_20-24_MRD_U15",
-                "PT_F_20-24_MRD_U18",
-                "PT_M_20-24_MRD_U18",
-                "PT_CHLD_5-17_LBR_ECON",
-                "PT_CHLD_5-17_LBR_ECON-HC",
-            ],
-            "default_graph": "bar",
-            "default": "PT_F_20-24_MRD_U15",
-        },
-        "AREA_3": {
-            "name": "Health and Nutrition",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "indicators": [
-                "HT_SH_ACS_UNHC",
-                "HT_ADOL_UNMETMED_NOUNMET",
-                "HT_ADOL_UNMETMED_TOOEXP",
-                "HT_ADOL_UNMETMED_TOOFAR",
-                "HT_ADOL_UNMETMED_WAITING",
-                "HT_ADOL_UNMETMED_TOOEFW",
-                "HT_ADOL_UNMETMED_NOTIME",
-                "HT_ADOL_UNMETMED_NOKNOW",
-                "HT_ADOL_UNMETMED_FEAR",
-                "HT_ADOL_UNMETMED_HOPING",
-                "HT_ADOL_UNMETMED_OTH",
-                "HT_SH_XPD_CHEX_GD_ZS",
-                "HT_SH_XPD_GHED_GD_ZS",
-                "HT_SH_XPD_GHED_GE_ZS",
-                "HT_SH_XPD_GHED_CH_ZS",
-                "HT_SH_XPD_GHED_PP_CD",
-                "HT_SH_XPD_GHED_PC_CD",
-                "HT_SH_XPD_PVTD_CH_ZS",
-                "HT_SH_XPD_PVTD_PP_CD",
-                "HT_SH_XPD_PVTD_PC_CD",
-                "HT_SH_XPD_CHEX_PP_CD",
-                "HT_SH_XPD_CHEX_PC_CD",
-                "HT_SH_XPD_OOPC_CH_ZS",
-                "HT_SH_XPD_OOPC_PP_CD",
-                "HT_SH_XPD_OOPC_PC_CD",
-                "HT_INS_COV",
-                "CME_MRY0T4",
-                "CME_MRY0",
-                "CME_MRM0",
-                "MNCH_MMR",
-                "MNCH_SAB",
-                "MNCH_PNCMOM",
-                "CME_TMY0T4",
-                "CME_SBR",
-                "CME_PND",
-                "HT_U5DEATH_AIDS",
-                "HT_U5DEATH_DIAR",
-                "HT_U5DEATH_PERT",
-                "HT_U5DEATH_TETA",
-                "HT_U5DEATH_MEAS",
-                "HT_U5DEATH_MENI",
-                "HT_U5DEATH_MALA",
-                "HT_U5DEATH_PNEU",
-                "HT_U5DEATH_PRET",
-                "HT_U5DEATH_INTR",
-                "HT_U5DEATH_SEPS",
-                "HT_U5DEATH_OTHE",
-                "HT_U5DEATH_CONG",
-                "HT_U5DEATH_NCDS",
-                "HT_U5DEATH_INJU",
-                "MNCH_CSEC",
-                "MNCH_PNCMOM",
-                "MNCH_ANC4",
-                "MNCH_PNEUCARE",
-                "IM_DTP3",
-                "HT_SH_ACS_DTP3",
-                "HT_SH_ACS_HPV",
-                "HT_SH_ACS_MCV2",
-                "HT_SH_ACS_PCV3",
-                "IM_MCV1",
-                "HT_DIST80DTP3_P",
-                "HT_COVERAGE_DTP3",
-                "NT_BW_LBW",
-                "NT_BF_EIBF",
-                "NT_BF_EXBF",
-                "NT_CF_MAD",
-                "NT_ANT_WHZ_PO2",
-                "NT_ANT_HAZ_NE2",
-                "HT_SN_STA_OVWGTN",
-                "HT_SH_STA_ANEM",
-                "HT_SH_STA_ANEM_NPRG",
-                "HT_SH_STA_ANEM_PREG",
-                "NT_BW_UNW",
-                "FT_SP_DYN_ADKL",
-                "HT_SH_FPL_MTMM",
-                "HT_ADOL_MT",
-                "HT_SH_PRV_SMOK",
-                "HT_CDRT_SELF_HARM",
-                "HT_ADOL_MEAL",
-                "HT_ADOL_NO_EXCS",
-                "HT_ADOL_VGRS_EXCS",
-                "HT_CHLD_BODY_RGHT",
-                "HT_CHLD_BODY_NO_RGHT",
-                "HT_CHLD_REG_SMOK",
-                "HT_CHLD_DRNK",
-                "HT_ADOL_HIGH_SATS",
-                "HT_ADOL_LOW_SATS",
-                "HT_SH_HIV_INCD",
-                "HVA_PMTCT_ARV_CVG",
-                "HT_SH_HIV_0014",
-                "HVA_EPI_LHIV_0-19",
-                "HVA_EPI_LHIV_15-24",
-                "HVA_EPI_INF_RT_0-14",
-                "HVA_EPI_INF_RT_10-19",
-                "HVA_PED_ART_NUM",
-                "HVA_PED_ART_CVG",
-                "WS_PPL_W-SM",
-                "WS_PPL_S-SM",
-                "WS_PPL_H-B",
-                "WS_PPS_S-OD",
-                "HT_NO_BTH_SHW_FLSH",
-            ],
-            "default_graph": "bar",
-            "default": "FT_SP_DYN_ADKL",
-        },
-        "AREA_4": {
-            "name": "Poverty and Social Protection",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "indicators": [
-                "PV_SI_POV_EMP1",
-                "PV_SI_POV_DAY1",
-                "PV_SI_POV_UMIC",
-                "PV_SDG_SI_POV_NAHC",
-                "PV_WB_SI_POV_NAHC",
-                "PV_AROPE",
-                "PV_AROPRT",
-                "PV_SD_MDP_CSMP",
-                "PV_SD_MDP_MUHHC",
-                "PV_SD_MDP_MUHC",
-                "PV_SI_POV_MDIM",
-                "PV_SI_POV_MDIM_17",
-                "WS_PPL_W-B",
-                "WS_PPL_S-B",
-                "PV_SEV_MAT_DPRT",
-                "PV_SI_COV_BENFTS",
-                "PV_SI_COV_LMKT",
-                "PV_SI_COV_SOCAST",
-                "PV_SI_COV_SOCINS",
-                "PV_SI_COV_WKINJRY",
-                "PV_SI_COV_CHLD",
-                "PV_SI_COV_DISAB",
-                "PV_SI_COV_MATNL",
-                "PV_SI_COV_POOR",
-                "PV_SI_COV_UEMP",
-                "PV_SI_COV_VULN",
-                "PV_SI_COV_PENSN",
-            ],
-            "default_graph": "bar",
-            "default": "PV_SI_POV_EMP1",
-        },
-        "AREA_5": {
-            "name": "Child Rights Landscape and Governance",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "indicators": [
-                "DM_CHLD_POP",
-                "DM_ADOL_POP",
-                "DM_CHLD_POP_PT",
-                "DM_CHLD_TWO_PRNT",
-                "DM_FRATE_TOT",
-                "DM_POP_URBN",
-                "EC_HDI",
-                "EC_NY_GDP_MKTP_KD_ZG",
-                "EC_NY_GDP_PCAP_KD_ZG",
-                "EC_NE_DAB_TOTL_ZS",
-                "EC_TEC_GRL_GOV_EXP",
-                "EC_TEC_CNT_GOV_EXP",
-                "EC_TEC_STA_GOV_EXP",
-                "EC_TEC_LOC_GOV_EXP",
-                "EC_TEC_SSF_EXP",
-                "EC_GR_G14_GDP",
-                "EDU_FIN_EXP_PT_GDP",
-                "HT_SH_XPD_GHED_GD_ZS",
-                "EC_SP_GOV_EXP_GDP",
-                "EC_SP_GOV_EXP_TOT",
-                "EC_NY_GDP_PCAP_PP_CD",
-                "EC_NY_GNP_ATLS_CD",
-                "EC_NY_GNP_PCAP_CD",
-                "EC_GC_DOD_TOTL_GD_ZS",
-                "EC_SI_POV_GINI",
-                "EC_SL_UEM_TOTL_NE_ZS",
-                "DM_POP_NETM",
-                "DM_SM_POP_REFG",
-                "DM_SM_POP_REFG_OR",
-                "DM_ASYL_FRST",
-                "DM_ASYL_UASC",
-                "MG_INTNL_MG_CNTRY_DEST_PS",
-                "MG_INTNL_MG_CNTRY_DEST_RT",
-                "PP_SG_NHR_IMPLN",
-                "PP_SG_NHR_INTEXSTN",
-                "PP_SG_NHR_NOSTUSN",
-                "PP_SG_NHR_NOAPPLN",
-                "JJ_CHLD_CRIME",
-                "JJ_CHLD_CRIMERT",
-                "CR_IQ_SCI_OVRL",
-                "CR_SG_STT_FPOS",
-                "CR_SG_STT_NSDSFND",
-                "CR_SG_STT_NSDSIMPL",
-                "CR_SG_STT_NSDSFDGVT",
-                "CR_SG_STT_NSDSFDDNR",
-                "CR_SG_STT_NSDSFDOTHR",
-                "CR_SG_STT_CAPTY",
-                "CR_SG_REG_CENSUSN",
-                "EDU_FIN_EXP_PT_GDP",
-                "HT_SH_XPD_GHED_GD_ZS",
-                "EC_SP_GOV_EXP_GDP",
-                "EC_TOT_PUB_EXP_GDP",
-                "EC_TOT_PUB_EXP_TOT",
-                "EC_FAM_PUB_EXP_GDP",
-                "EC_FAM_PUB_EXP_TOT",
-                "HT_SH_XPD_CHEX_GD_ZS",
-                "HT_SH_XPD_GHED_GD_ZS",
-                "HT_SH_XPD_CHEX_PP_CD",
-                "EDU_FIN_EXP_PT_TOT",
-                "EDU_FIN_EXP_CONST_PPP",
-            ],
-            "default_graph": "bar",
-            "default": "DM_CHLD_POP_PT",
-        },
-        "AREA_6": {
-            "name": "Participation and Civil Rights",
-            "graphs": {
-                "bar": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        barmode="group",
-                        # text="TIME_PERIOD",
-                        text="OBS_VALUE",
-                        hover_name="Country_name",
-                    ),
-                },
-                "line": {
-                    "options": dict(
-                        x="TIME_PERIOD",
-                        y="OBS_VALUE",
-                        color="Country_name",
-                        hover_name="Country_name",
-                        line_shape="spline",
-                        render_mode="svg",
-                    ),
-                    "trace_options": dict(mode="lines+markers"),
-                },
-            },
-            "indicators": [
-                "PT_CHLD_Y0T4_REG",
-                "PP_SG_REG_BRTH90N",
-                "PP_SG_REG_DETH75N",
-                "PP_IT_USE_ii99",
-                "PP_SE_ADT_ACTS_ATCH",
-                "PP_SE_ADT_ACTS_CPT",
-                "PP_SE_ADT_ACTS_CDV",
-                "PP_SE_ADT_ACTS_SSHT",
-                "PP_SE_ADT_ACTS_PRGM",
-                "PP_SE_ADT_ACTS_PST",
-                "PP_SE_ADT_ACTS_SFWR",
-                "PP_SE_ADT_ACTS_TRFF",
-                "PP_SE_ADT_ACTS_CMFL",
-                "PP_IT_MOB_OWN",
-                "PP_ADOL_TVGM",
-                "PP_ADOL_INET",
-                "PP_ADOL_ITXT",
-                "PP_ADOL_WORK_PAID",
-                "PP_ADOL_WORK_HOME",
-            ],
-            "default_graph": "bar",
-            "default": "PT_CHLD_Y0T4_REG",
-        },
-    },
-}
-
-main_title = "Country Snapshot"
-
-
-def get_layout(**kwargs):
-    kwargs["indicators"] = indicators_dict
-    kwargs["main_title"] = main_title
-    kwargs["is_country_profile"] = True
-    return get_base_layout(**kwargs)
