+++ transmonee_dashboard/src/transmonee_dashboard/pages/data_query.py
@@ -20,11 +20,11 @@ from . import (
     countries_iso3_dict,
     data,
     years,
-    df_topics_subtopics,
+    # df_topics_subtopics,
     topics_subtopics,
     data_sources,
-    df_sources,
-    indicators_disagg_details,
+    # df_sources,
+    # indicators_disagg_details,
     get_search_countries,
 )
